requirements.txt
requests
python-decouple
uvicorn
fastapi
fastapi-sessions
starlette